MoneyTransaction/requirements.txt
absl-py==0.13.0
altair==4.1.0
aniso8601==9.0.1
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
colorama==0.4.4
cryptography==37.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-dictionaryfield==1.0
django-filter==21.1
django-jsonfield==1.4.1
django-jsonfield-backport==1.0.5
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drfpasswordless==1.5.8
EcoNameTranslator==2.1
entrypoints==0.3
export==0.1.2
filelock==3.4.2
Flask==2.0.1
Flask-RESTful==0.3.9
flatbuffers==1.12
fonttools==4.28.3
gast==0.4.0
geographiclib==1.52
geojson==2.5.0
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.40.0
h5py==3.1.0
identify==2.4.12
idna==3.2
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.1.0
jsonfield==1.0.3
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.6
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
phonenumbers==8.12.31
pickleshare==0.7.5
pika==1.2.0
Pillow==8.3.2
platformdirs==2.4.1
pre-commit==2.18.1
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.18.0
psycopg2==2.9.3
py4j==0.10.9.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
PyJWT==2.3.0
Pympler==0.9
pyowm==3.2.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pyspark==3.2.0
python-dateutil==2.8.2
python-decouple==3.5
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
source==1.2.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
streamlit==1.2.0
taxon-parser==0.2.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
validators==0.18.2
virtualenv==20.13.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
windows-curses==2.3.0
wrapt==1.12.1
xlwt==1.3.0